change app to user as recording_limit quantity of time stored on the project instead of recording count
Changes
 - added getUsage route that return qty of minutes stored on a project
 - added storage usage info to the project dashboard
 - added `upload_time` field to recordings table and update `file_size` field to
   type float to start monitoring when users uploaded each recording
 - added `state` and `duration` fields to upload_processing
 - updated uploads router to use minutes usage, and verify site existence
 - created utils/uploader.js to with deal with upload processing logic and now
   file_size is given by fs.stat
 - uploadQueue only exports method enqueue
 - rewrote unit tests for uploads (uploader.js, routes/uploads.js, upload-queue.js)

other changes
 - refactor classificationDelete with async to remove code duplication and
   improve readability
 - bug fix - recordings.delete recordings thumbnails were not being deleted
   with audio file
 - create rob-the-builder.js dummy constructor builder for testing

+++ routes/uploads.js
@@ -11,7 +11,7 @@ var model = require('../model');
 var audioTools= require('../utils/audiotool');
 var tmpFileCache = require('../utils/tmpfilecache');
 var formatParse = require('../utils/format-parse');
-var uploadQueue = require('../utils/upload-queue').uploadQueue;
+var uploadQueue = require('../utils/upload-queue');
 
 var deleteFile = function(filename) {
     fs.unlink(filename, function(err) {
@@ -21,8 +21,6 @@ var deleteFile = function(filename) {
 
 // middleware to handle upload auth
 var authorize = function(req, res, next) {
-    req.upload = {};
-    
     if(req.session && req.session.loggedIn) { 
         if(!req.query.project || !req.query.site || !req.query.nameformat) {
             return res.status(400).json({ error: "missing parameters" });
@@ -45,37 +43,53 @@ var authorize = function(req, res, next) {
         
         req.upload = {
             userId: req.session.user.id,
-            projectId: req.query.project,
-            siteId: req.query.site,
+            projectId: Number(req.query.project),
+            siteId: Number(req.query.site),
             nameFormat: req.query.nameformat,
         };
         
         next();
     }
+    
     // verify token
     else if(req.token) {
-        model.sites.findById(req.token.site, function(err, rows) {
-            if(err) return next(err);
-            if(!rows.length) return res.sendStatus(401);
-            
-            var site = rows[0];
-            
-            if(site.token_created_on !== req.token.iat) return res.sendStatus(401);
-            
-            req.upload = {
-                userId: 0,
-                projectId: req.token.project,
-                siteId: req.token.site,
-                nameFormat: "Arbimon",
-            };
-            
-            next();
-        });
+        
+        req.upload = {
+            userId: 0,
+            projectId: Number(req.token.project),
+            siteId: Number(req.token.site),
+            nameFormat: "Arbimon",
+        };
+        
+        next();
     }
     else {
         // error not logged user nor site token
         return res.sendStatus(401);
     }
+};
+
+var verifySite = function(req, res, next) {
+    model.sites.findById(req.upload.siteId, function(err, rows) {
+        if(err) return next(err);
+        
+        if(!rows.length) {
+            return res.status(403).json({ error: "site does not exist" });
+        }
+        
+        var site = rows[0];
+        
+        // verify token is valid to the system
+        if(req.token && (site.token_created_on !== req.token.iat)) {
+            return res.status(400).json({ error: "can not use revoked token" });
+        }
+        
+        if(site.project_id !== req.upload.projectId) {
+            return res.status(403).json({ error: "site does not belong to project"});
+        }
+        
+        next();
+    });
 };
 
 var receiveUpload = function(req, res, next) {
@@ -84,24 +98,27 @@ var receiveUpload = function(req, res, next) {
         getProjectInfo: function(callback) {
             model.projects.findById(req.upload.projectId, callback);
         },
-        getTotal: function(callback) {
-            model.projects.totalRecordings(req.upload.projectId, callback);
+        getUsage: function(callback) {
+            model.projects.getStorageUsage(req.upload.projectId, callback);
         }
     }, 
     function(err, results) {
         if(err) return next(err);
         
         var project = results.getProjectInfo[0][0];
-        var total = results.getTotal[0][0].count;
+        var minsUsage = results.getUsage[0][0].min_usage;
         
-        if(total >= project.recording_limit) {
+        debug('limit', project.recording_limit);
+        debug('usage', minsUsage);
+        
+        if(minsUsage >= project.recording_limit) {
             return res.status(401).json({ error: "Project Recording limit reached"});
         }
         
         var upload = {};
         var error = false;
         
-        if(!req.busboy) return res.sendStatus(400);
+        if(!req.busboy) return res.status(400).json({ error: "no data" });
         
         req.busboy.on('field', function(fieldname, val) {
             debug('field %s = %s', fieldname, val);
@@ -109,16 +126,19 @@ var receiveUpload = function(req, res, next) {
             if(fieldname === 'info') {
                 try {
                     upload.metadata = JSON.parse(val);
-                    
-                    if(!upload.metadata.recorder || 
-                        !upload.metadata.sver ||
-                        !upload.metadata.mic) {
-                        return res.status(400).json({ error: "missing basic metadata" });
-                    }
                 }
                 catch(err) {
                     error = true;
                     return res.status(400).json({ error: err.message });
+                }
+                
+                var notValid = !upload.metadata.recorder || 
+                                !upload.metadata.sver ||
+                                !upload.metadata.mic;
+                    
+                if(notValid && !error) {
+                    error = true;
+                    return res.status(400).json({ error: "missing basic metadata" });
                 }
             }
         });
@@ -156,7 +176,7 @@ var receiveUpload = function(req, res, next) {
             debug('filename: ', upload.name);
             
             async.waterfall([
-                function checkExists(callback) {
+                function recNotExists(callback) {
                     model.recordings.exists({
                         site_id: req.upload.siteId,
                         filename: upload.FFI.filename
@@ -172,13 +192,12 @@ var receiveUpload = function(req, res, next) {
                         }
                         
                         callback();
-                        
-                        
                     });
                 },
                 function getAudioInfo(callback) {
                     audioTools.info(upload.path, function(code, info) {
                         if(code !== 0) {
+                            deleteFile(upload.path);
                             return res.status(500).json({ error: "error getting audio file info" });
                         }
                         
@@ -192,14 +211,26 @@ var receiveUpload = function(req, res, next) {
                     upload.userId = req.upload.userId;
                     upload.info = info;
                     
-                    
                     if(info.duration >= 3600) {
-                        return res.status(403).json({ error: "recording is too long, please contact support" });
-                    } 
-                    else if(info.duration >= 120) {
+                        deleteFile(upload.path);
+                        return res.status(403).json({ error: "Recording is too long, please contact support" });
+                    }
+                    
+                    debug('new usage', (info.duration/60)+minsUsage);
+                    
+                    if((info.duration/60)+minsUsage > project.recording_limit) {
+                        deleteFile(upload.path);
+                        var msg = "Recording is too long, there is only " + Math.round((project.recording_limit-minsUsage)*100)/100 +
+                            " minutes of space left";
+                        return res.status(401).json({ error: msg });
+                    }
+                    
+                    if(info.duration > 120) {
                         audioTools.splitter(upload.path, info.duration, function(err, files) {
+                            if(err) return next(err);
+                            
                             var i = 0;
-                            async.eachSeries(files, function(f, callback) {
+                            async.eachSeries(files, function(f, nextUpload) {
                                 var uploadPart = _.cloneDeep(upload);
                                 
                                 uploadPart.FFI.filename = uploadPart.FFI.filename +'.p'+ (i+1);
@@ -210,20 +241,27 @@ var receiveUpload = function(req, res, next) {
                                 uploadPart.path = f;
                                 
                                 debug('upload', uploadPart);
-                                uploadQueue.push(_.cloneDeep(uploadPart));
+                                uploadQueue.enqueue(_.cloneDeep(uploadPart), function(err) {
+                                    if(err) return nextUpload(err);
+                                    
+                                    i++;
+                                    nextUpload();
+                                });
+                            }, function splitDone(err2) {
+                                if(err2) return next(err2);
                                 
-                                i++;
-                                callback();
-                            }, function() {
                                 deleteFile(upload.path);
-                                res.status(202).send("upload done!");
+                                res.status(202).json({ success: "upload done!" });
                             });
                         });
                     }
                     else {
                         debug('upload', upload);
-                        uploadQueue.push(upload);
-                        res.status(202).send("upload done!");
+                        uploadQueue.enqueue(upload, function(err) {
+                            if(err) return next(err);
+                            
+                            res.status(202).json({ success: "upload done!" });
+                        });
                     }
                 }
             ]);
@@ -234,6 +272,6 @@ var receiveUpload = function(req, res, next) {
     });
 };
 
-router.post('/audio', authorize, receiveUpload);
+router.post('/audio', authorize, verifySite, receiveUpload);
 
 module.exports = router;
